+++ src/data/movie/MovieSet.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "data/TmdbId.h"
+#include "utils/Meta.h"
 
 #include <QString>
 
@@ -14,4 +15,9 @@ struct MovieSet
     TmdbId tmdbId{TmdbId::NoId};
     QString name;
     QString overview;
+
+    /// \brief Returns this collection under a new name, keeping overview and id.
+    /// \details Only for renaming a collection.  Moving a movie to a _different_
+    ///          collection must not use this: overview and id describe the old one.
+    ELCH_NODISCARD MovieSet renamedTo(QString newName) const;
 };